+++ src/bookmark_reader/firefox.rs
@@ -127,6 +127,7 @@ impl Firefox {
 }
 
 /// A bookmark reader to read bookmarks in JSON format from Firefox.
+#[derive(Debug)]
 pub struct FirefoxBookmarkReader;
 
 impl ReadBookmark for FirefoxBookmarkReader {
@@ -204,6 +205,7 @@ impl ReadBookmark for FirefoxBookmarkReader {
 }
 
 /// A bookmark reader to read compressed bookmarks in JSON format from Firefox.
+#[derive(Debug)]
 pub struct FirefoxCompressedBookmarkReader;
 
 impl ReadBookmark for FirefoxCompressedBookmarkReader {
@@ -332,7 +334,7 @@ mod tests {
         let mut bookmark_file = Cursor::new(decompressed_bookmarks);
 
         let raw_bookmarks = bookmark_reader.read(&mut bookmark_file).unwrap();
-        let mut source_bookmarks = SourceBookmarks::new();
+        let mut source_bookmarks = SourceBookmarks::default();
         let source = Source::new("dummy_path", vec![]);
 
         let res = bookmark_reader.parse(&raw_bookmarks, &source, &mut source_bookmarks);
@@ -355,7 +357,7 @@ mod tests {
         let mut bookmark_file = Cursor::new(compressed_bookmarks);
 
         let raw_bookmarks = bookmark_reader.read(&mut bookmark_file).unwrap();
-        let mut source_bookmarks = SourceBookmarks::new();
+        let mut source_bookmarks = SourceBookmarks::default();
         let source = Source::new("dummy_path", vec![]);
 
         let res = bookmark_reader.parse(&raw_bookmarks, &source, &mut source_bookmarks);
@@ -377,7 +379,7 @@ mod tests {
         let mut bookmark_file = Cursor::new(decompressed_bookmarks);
 
         let raw_bookmarks = bookmark_reader.read(&mut bookmark_file).unwrap();
-        let mut source_bookmarks = SourceBookmarks::new();
+        let mut source_bookmarks = SourceBookmarks::default();
         let source = Source::new("dummy_path", vec![String::from("dev")]);
 
         let res = bookmark_reader.parse(&raw_bookmarks, &source, &mut source_bookmarks);
@@ -401,7 +403,7 @@ mod tests {
         let mut bookmark_file = Cursor::new(compressed_bookmarks);
 
         let raw_bookmarks = bookmark_reader.read(&mut bookmark_file).unwrap();
-        let mut source_bookmarks = SourceBookmarks::new();
+        let mut source_bookmarks = SourceBookmarks::default();
         let source = Source::new("dummy_path", vec![String::from("dev")]);
 
         let res = bookmark_reader.parse(&raw_bookmarks, &source, &mut source_bookmarks);