+++ tests/test_update.rs
@@ -30,7 +30,7 @@ async fn test_update() {
     cmd.args(["config", "--source", source.to_str().unwrap()]);
     cmd.output().unwrap();
 
-    let bookmarks = common::test_bookmarks(&temp_dir);
+    let bookmarks = common::test_bookmarks(temp_path);
     assert!(bookmarks.is_empty());
 
     println!("Execute 'bogrep import'");
@@ -39,7 +39,7 @@ async fn test_update() {
     cmd.args(["import"]);
     cmd.output().unwrap();
 
-    let bookmarks = common::test_bookmarks(&temp_dir);
+    let bookmarks = common::test_bookmarks(temp_path);
     assert_eq!(bookmarks.len(), 3);
 
     println!("Execute 'bogrep fetch'");
@@ -48,7 +48,7 @@ async fn test_update() {
     cmd.args(["fetch"]);
     cmd.output().unwrap();
 
-    let bookmarks = common::test_bookmarks(&temp_dir);
+    let bookmarks = common::test_bookmarks(temp_path);
     assert_eq!(bookmarks.len(), 3);
     for bookmark in &bookmarks {
         assert!(bookmark.last_cached.is_some());
@@ -60,7 +60,7 @@ async fn test_update() {
     cmd.args(["update"]);
     cmd.output().unwrap();
 
-    let bookmarks = common::test_bookmarks(&temp_dir);
+    let bookmarks = common::test_bookmarks(temp_path);
     assert_eq!(bookmarks.len(), 3);
     for bookmark in &bookmarks {
         assert!(bookmark.last_cached.is_some());